% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_help.R
\name{clockwise_angle}
\alias{clockwise_angle}
\title{Clockwise angle}
\usage{
clockwise_angle(a, b)
}
\arguments{
\item{a}{vector}

\item{b}{vector}
}
\value{
clockwise angle between vector a and b
}
\description{
Calculates clockwise angle between two vectors.
Positive means clockwise change and negative counter clockwise.
}
