% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_processing_tracked.R
\name{process_data_dt}
\alias{process_data_dt}
\title{Title}
\usage{
process_data_dt(
  input_dir,
  output_dir,
  save_as,
  radius,
  drop_contours,
  frame_rate
)
}
\arguments{
\item{input_dir}{the directory containing the tracked data}

\item{output_dir}{where the analysis should be saved}

\item{save_as}{the folder name under which the analysis should be saved}

\item{radius}{the radius of the petri dish in mm}

\item{drop_contours}{if contours should be dropped (contours will be useful for the animation of larvae)}

\item{frame_rate}{the frame rate of the tracked videos}
}
\value{
data.frame with the analyzed data set
}
\description{
Title
}
